RealtyHub/RealtyHubBackend/src/main/java/org/example/business/impl/LoginManager.java
package org.example.business.impl;


import lombok.RequiredArgsConstructor;
import org.example.business.AccessTokenEncoder;
import org.example.business.ILoginManager;
import org.example.business.exception.InvalidCredentialsException;
import org.example.domain.LoginRequest;
import org.example.domain.LoginResponse;
import org.example.domain.classes.AccessToken;
import org.example.persistance.UserRepository;
import org.example.persistance.entity.UserEntity;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class LoginManager implements ILoginManager {
    private final UserRepository userRepository;
    private final PasswordEncoder passwordEncoder;
    private final AccessTokenEncoder accessTokenEncoder;

    @Override
    public LoginResponse login(LoginRequest loginRequest) {
        UserEntity user = userRepository.findByUsername(loginRequest.getUsername());
        if (user == null) {
            throw new InvalidCredentialsException();
        }

        if (!matchesPassword(loginRequest.getPassword(), user.getPassword())) {
            throw new InvalidCredentialsException();
        }

        String accessToken = generateAccessToken(user);
        return LoginResponse.builder().accessToken(accessToken).build();
    }

    private boolean matchesPassword(String rawPassword, String encodedPassword) {
        return passwordEncoder.matches(rawPassword, encodedPassword);
    }

    private String generateAccessToken(UserEntity user) {
        Long userID = user.getUserinfo() != null ? user.getUserinfo().getId() : null;
        List<String> roles = user.getUserRoles().stream()
                .map(userRole -> userRole.getRole().toString())
                .toList();

        return accessTokenEncoder.encode(
                AccessToken.builder()
                        .subject(user.getUsername())
                        .roles(roles)
                        .userId(userID)
                        .build());
    }

}
